requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
asgiref==3.3.1
    # via django
django-crum==0.7.9
    # via -r requirements/base.in
django==3.1.5
    # via
    #   -r requirements/base.in
    #   django-crum
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements/base.in
pytz==2020.5
    # via django
sqlparse==0.4.1
    # via django
